requirements.txt
discord.py == 2.3.2
pyjokes == 0.6.0
requests == 2.31.0